chore(pkg/plugins/runtime/k8s) move utility function to util package
